Refactor model packages and update recommendation API
Move domain models into org.tvl.tvlooker.domain.model.entity and update imports accordingly; relocate RecommendationContext to domain.motor.utils. Introduce ScoredItem data structure and AggregationStrategy interface. Replace EvaluationStrategy with aggregation-focused API: RecommendationEngine now returns List<ScoredItem> and exposes setStrategies/setAggregationStrategy; add HybridRecommendationEngine stub implementing the new API. Update DataProvider to use the new RecommendationContext type. Also update .gitignore to exclude .agents/ and skills-lock.json.
