+++ src/main/java/org/tvl/tvlooker/domain/data_provider/DataProvider.java
@@ -1,6 +1,6 @@
 package org.tvl.tvlooker.domain.data_provider;
 
-import org.tvl.tvlooker.domain.model.RecommendationContext;
+import org.tvl.tvlooker.domain.motor.utils.RecommendationContext;
 
 public interface DataProvider <T>{
     T provide(RecommendationContext context);
