+++ src/main/java/org/tvl/tvlooker/domain/model/entity/Interaction.java
@@ -1,4 +1,4 @@
-package org.tvl.tvlooker.domain.model;
+package org.tvl.tvlooker.domain.model.entity;
 
 import java.sql.Timestamp;
