+++ src/main/java/org/tvl/tvlooker/domain/motor/HybridRecommendationEngine.java
@@ -1,0 +1,47 @@
+package org.tvl.tvlooker.domain.motor;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.stereotype.Component;
+import org.tvl.tvlooker.domain.data_structure.ScoredItem;
+import org.tvl.tvlooker.domain.motor.utils.RecommendationContext;
+import org.tvl.tvlooker.domain.model.entity.User;
+import org.tvl.tvlooker.domain.strategy.aggregation_strategy.AggregationStrategy;
+import org.tvl.tvlooker.domain.strategy.recommendation_strategy.RecommendationStrategy;
+
+import java.util.List;
+
+/**
+ * The HybridRecommendationEngine class is a concrete implementation of the RecommendationEngine interface that combines
+ * multiple recommendation strategies and an evaluation strategy to generate item recommendations for users. It allows
+ * for flexible configuration of strategies and can be extended to include various recommendation algorithms.
+ */
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
+@Component
+public class HybridRecommendationEngine implements RecommendationEngine {
+    /** A list of recommendation strategies that the engine will use to generate recommendations. */
+    private  List<RecommendationStrategy> strategies;
+    /** An evaluation strategy that the engine will use to evaluate the effectiveness of the recommendations. */
+    private AggregationStrategy aggregationStrategy;
+
+    /** Generates a list of scored items as recommendations for a given user based on the provided recommendation
+     * context. This method combines the results from multiple recommendation strategies and applies an aggregation
+     * strategy to produce the final list of recommendations. The implementation can be extended to include specific
+     * logic for generating recommendations based on user interactions, preferences, and other relevant data.
+     *
+     * @param user    The user for whom the recommendations are being generated.
+     * @param context The recommendation context that contains relevant data such as users, items, interactions, and
+     *                registered data providers.
+     * @return A list of scored items representing the recommended items for the user, along with their scores and
+     *         explanations.
+     */
+    @Override
+    public List<ScoredItem> recommend(User user, RecommendationContext context) {
+        return List.of();
+    }
+}
