+++ src/main/java/org/tvl/tvlooker/domain/motor/RecommendationEngine.java
@@ -2,11 +2,10 @@ package org.tvl.tvlooker.domain.motor;
 
 import java.util.List;
 
-import org.tvl.tvlooker.domain.model.Interaction;
-import org.tvl.tvlooker.domain.model.Item;
-import org.tvl.tvlooker.domain.model.RecommendationContext;
-import org.tvl.tvlooker.domain.model.User;
-import org.tvl.tvlooker.domain.strategy.evaluation_strategy.EvaluationStrategy;
+import org.tvl.tvlooker.domain.data_structure.ScoredItem;
+import org.tvl.tvlooker.domain.motor.utils.RecommendationContext;
+import org.tvl.tvlooker.domain.model.entity.User;
+import org.tvl.tvlooker.domain.strategy.aggregation_strategy.AggregationStrategy;
 import org.tvl.tvlooker.domain.strategy.recommendation_strategy.RecommendationStrategy;
 
 /**
@@ -15,8 +14,7 @@ import org.tvl.tvlooker.domain.strategy.recommendation_strategy.RecommendationSt
  * items, setting recommendation strategies, and setting an evaluation strategy.
  */
 public interface RecommendationEngine {
-    List<Item> recommend(User user, List<Interaction> interactions, RecommendationContext context);
-    void setRecommendationStrategies(List<RecommendationStrategy> strategies);
-    void setEvaluationStrategy(EvaluationStrategy evaluationStrategy);
-    
+    List<ScoredItem> recommend(User user, RecommendationContext context);
+    void setStrategies(List<RecommendationStrategy> strategies);
+    void setAggregationStrategy(AggregationStrategy aggregationStrategy);
 }
