+++ src/main/java/org/tvl/tvlooker/domain/strategy/aggregation_strategy/AggregationStrategy.java
@@ -1,0 +1,5 @@
+package org.tvl.tvlooker.domain.strategy.aggregation_strategy;
+
+public interface AggregationStrategy {
+
+}
