--- src/main/java/org/tvl/tvlooker/domain/strategy/evaluation_strategy/EvaluationStrategy.java
@@ -1,5 +1,0 @@
-package org.tvl.tvlooker.domain.strategy.evaluation_strategy;
-
-public interface EvaluationStrategy {
-
-}
